requirements.txt
dataclasses==0.6
Django==2.2.3
django-rest-framework-mongoengine==3.3.1
djangorestframework==3.9.4
djangotoolbox==1.8.0
gunicorn==19.9.0
mongoengine==0.9.0
pkg-resources==0.0.0
pymongo==3.8.0
pytz==2019.1
six==1.12.0
sqlparse==0.2.4
